feat: Allow for multiple proto compilers (#14)
The go gazelle plugin has a go_proto_compilers directive which accepts a
list of go_proto_compiler targets,
while currently the swift gazelle plugin has a swift_proto_compiler
directive which can be set multiple times with different flavors. These
flavors are: swift_proto (SwiftProtobuf), and swift_client_proto /
swift_server_proto (grpc-swift).

Swift needs these different flavors because native apps are often
concerned with app size and only need to act as a client, while on the
backend, services might act as both clients and servers and are not
concerned with binary bloat.

The current directive is limiting, however, for consumers that want to
run multiple proto compilers, e.g. for generating ORM bindings or other
code in addition to regular protos.

For better parity with the Go gazelle plugin, this PR renames the
directive to swift_proto_compilers and makes it support a list of
compilers for each flavor. Additionally, if the flavor is omitted, the
flavor is inferred to be swift_proto.